+++ packages/.vitepress/shared/unocss.theme.ts
@@ -1200,11 +1200,11 @@ export const easing = {
 export const transitionProperty = {
   none: "none",
   all: "all",
-  colors: "color,background-color,border-color,outline-color,text-decoration-color,fill,stroke",
+  colors: "color,background-color,border-color,text-decoration-color,fill,stroke",
   opacity: "opacity",
   shadow: "box-shadow",
   transform: "transform",
-  DEFAULT: "color,background-color,border-color,outline-color,text-decoration-color,fill,stroke,opacity,box-shadow,transform,filter,backdrop-filter"
+  DEFAULT: "color,background-color,border-color,text-decoration-color,fill,stroke,opacity,box-shadow,transform,filter,backdrop-filter"
 };
 export const lineWidth = {
   DEFAULT: "1px",